finish all tests
